Correct $itempos for repeats and $skip.
$itempos formerly reported an $itempos{from} that was the left side of
the skip before a $item.  Now $itempos{from} reports the actual
position of the left side of the $item after the $skip is processed.
Updated t/basics.t to reflect this.

$itempos formerly reported $itempos{from} for a repeating construct as
if the {from} were from the last repetition.  Now $itempos contains
the {from} from the leftmost non-$skip member of the repetition, and
the {to} of the rightmost member of the repetition.