Don't overwrite %item with repeated subrule matches in a production.
*** NON-BACKWARDS COMPATIBLE CHANGE! *** A repetition directive
such as 'id(s /,/)' correctly creates a temporary @item variable
to hold the 'id's that are matched.  That @item variable is them
used to set the real $item[] entry for that repetition.  The
same treatment is now given to %item.  Formerly, in a production like:

    id ',' id(s /,/)

matched against:

    xxx, yyy, zzz

The $item{id} entry which should be 'xxx' is overwritten by
'yyy' and then 'zzz' prior to the action being executed.  Now
'yyy' and 'zzz' set $item{id}, but in the private %item, which
goes out of scope once the repetition match completes.