Fix a circular reference memory leak. (RT.cpan.org #53710)
The use of $AUTOLOAD in sub AUTOLOAD{} in the generated parser's
namespace causes a circular reference, documented in perl5 RT #110248.
Applied the workaround documented there, to use ${"AUTOLOAD"} instead
of $AUTOLOAD.  This forces a GLOB lookup, which takes slightly longer,
at the expense of not leaking memory.  Since the GLOB lookup only
happens on "external" calls into the parser, this shouldn't be a big
performance hit for most grammars, unless they are called VERY often.